move MappingGraph class from Sofa.Component.LinearSystem to Sofa.Simulation.Core
