+++ Sofa/framework/Simulation/Core/test/CMakeLists.txt
@@ -4,6 +4,7 @@ project(Sofa.Simulation.Core_test)
 
 set(SOURCE_FILES
     Colors_test.cpp
+    MappingGraph_test.cpp
     ParallelForEach_test.cpp
     RequiredPlugin_test.cpp
     SceneCheckRegistry_test.cpp
@@ -16,6 +17,9 @@ set(SOURCE_FILES
     )
 
 add_executable(${PROJECT_NAME} ${SOURCE_FILES})
-target_link_libraries(${PROJECT_NAME} Sofa.Testing Sofa.Simulation.Core Sofa.LinearAlgebra.Testing)
+target_link_libraries(${PROJECT_NAME} Sofa.Testing Sofa.Simulation.Core Sofa.LinearAlgebra.Testing
+    Sofa.Component.StateContainer
+    Sofa.Component.Mapping.Linear
+)
 
 add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME})
